requirements.txt
python-dotenv
fastapi 
uvicorn[standard] 
pydantic 
sqlalchemy 
aiosqlite 
jinja2
openai
pandas
torch # for embedding FAQ
azure-core
